Patch submitted by bitbucket user scottawilliams71 with the following comment:
A known bug in jdMenu 1.4.1 is causing undesirable behavior when
clicking root level menu items. Menu items with sub menus work as
expected. See
http://forum.jquery.com/topic/jquery-root-menu-item-causes-menu-to-disappear-in-jdmenu-1-4-1
for discussion and JS update necessary to fix the issue. Since this
discussion dates to June/July 2008 and 1.4.1 is still latest release
of jdMenu, I would suggest fixing the code in the copy of
jquery.jdMenu.js that is distributed with tgext.menu. See attached
patch.