Adding test case for following bug report: --------------------------------------------------------------------------- I have notices the following behaviour, even in the latest release.
say I have 2 controller decorated like so :

   @navbar('Terminals || Manage', extras={'id':'terminals_manage'})
   @expose()
   def manage(self):
       return dict()

   @navbar('Terminals || Groups', extras={'id':'terminals_groups'})
   @expose()
   def group(self):
       return dict()

I will generate a menu with id like this

<ul id="navbar>
  <li id="terminals_manage">
       <ul>
           <li id="terminals_manage">


that is the top level menu has the same id. not a good thing to have 2 elements on the page with the same id.

the top element isn"'t used in any decorator.
---------------------------------------------------------------------------

The test case being added modifies the root controller in
tgext/menu/test/controllers/root.py to add id attributes to methods
bar and baz. Rather than using terminals_manage and terminals_groups,
I choose foo_1 and foo_2 as the id values.

I then added what their respective output values *should* be to the
output strings in tgext/menu/test/test_controller.py, and ran the
tests. The tests show output that looks as follows:

------------------
            <li id="foo_2"><a href="/bar">Bar</a></li>
            <li class="last" id="foo_1"><a href="/foo">Foo</a></li>
------------------

So, the data looks correct. Must return this bug to the user, and
request a test case that shows this misbehavior.